+++ src/App.tsx
@@ -1,0 +1,15 @@
+import StopWatch from "./StopWatch";
+import Timer from "./Timer";
+
+export default function App() {
+
+  return (
+    <main>
+      <StopWatch/>
+
+      <br/><br/>
+
+      <Timer/>
+    </main>
+  )
+}
