+++ src/main.tsx
@@ -1,10 +1,9 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import { CodeReview } from './index.tsx'
-import './index.css'
+import App from './App'
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
-    <CodeReview />
+    <App/>
   </React.StrictMode>,
 )
